requirements.txt
streamlit
langchain
langchain-community
sentence-transformers
faiss-cpu
neo4j
pandas
numpy
openai
python-dotenv
transformers
torch
scikit-learn
networkx
matplotlib